+++ views/ingredients.ejs
@@ -8,7 +8,7 @@
         <% } %>
     </ul>
   </div>
-  <form class="w-100" method="POST" action="/clearIngredients">
+  <form class="w-100" method="POST" action="/home/clearIngredients">
     <button class="btn btn-danger">Clear Ingredients</button>
 </form>
 <% } %>
